feat: create Auctions list component
